Move videos parts out to their own components
New: VideosGrid and VideosGridItem

and add no thumbnail found msg
